# src/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Kubernetes Namespace Provisioner API
  description: |
    TypeScript implementation of the Eclipse Che server's Kubernetes namespace provisioning API.
    
    This API provides endpoints for provisioning and managing Kubernetes namespaces where 
    authenticated users can create workspaces.
    
    ## Original Java Implementation
    This API is based on the Eclipse Che server endpoint:
    - Java Class: `org.eclipse.che.workspace.infrastructure.kubernetes.api.server.KubernetesNamespaceService`
    - Endpoint: `POST /kubernetes/namespace/provision`
    
    ## Authentication
    The API supports two authentication methods:
    1. **Bearer Token**: Format `userid:username` (e.g., `Bearer user123:johndoe`)
    2. **Basic Auth**: Username and userid as password (e.g., `johndoe:user123`)
    
    ## Features
    - ✅ Provision Kubernetes namespaces with configurable name templates
    - ✅ List all managed namespaces
    - ✅ Automatic namespace creation if not exists
    - ✅ Namespace metadata with status and attributes
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: Eclipse Public License 2.0
    url: https://www.eclipse.org/legal/epl-2.0/
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /kubernetes/namespace:
    get:
      tags:
        - kubernetes-namespace
      summary: List namespaces
      description: |
        Get k8s namespaces where user is able to create workspaces. 
        This operation can be performed only by authorized user.
        
        **Note**: This is a beta feature that may be significantly changed.
      operationId: listNamespaces
      security:
        - BearerAuth: []
        - BasicAuth: []
      responses:
        '200':
          description: The namespaces successfully fetched
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KubernetesNamespaceMeta'
              example:
                - name: che-johndoe
                  attributes:
                    phase: Active
                    default: 'true'
                - name: che-janedoe
                  attributes:
                    phase: Active
                    default: 'false'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /kubernetes/namespace/provision:
    post:
      tags:
        - kubernetes-namespace
      summary: Provision namespace
      description: |
        Provision k8s namespace where user is able to create workspaces. 
        This operation can be performed only by an authorized user.
        
        The namespace name is generated based on a configurable template. 
        By default, it uses the pattern `che-<username>` where `<username>` 
        is replaced with the authenticated user's username.
        
        If the namespace already exists, it returns the existing namespace metadata.
        
        **Note**: This is a beta feature that may be significantly changed.
        
        ## How it works
        1. Extracts user information from authentication context
        2. Evaluates namespace name using the configured template
        3. Creates namespace in Kubernetes if it doesn't exist
        4. Configures namespace with appropriate labels and annotations
        5. Returns namespace metadata including name and attributes
        
        ## Namespace Attributes
        - `phase`: Current namespace status (Active, Terminating, etc.)
        - `default`: Whether this is the user's default namespace (true/false)
      operationId: provisionNamespace
      security:
        - BearerAuth: []
        - BasicAuth: []
      responses:
        '200':
          description: The namespace successfully provisioned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KubernetesNamespaceMeta'
              example:
                name: che-johndoe
                attributes:
                  phase: Active
                  default: 'true'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /oauth:
    get:
      tags:
        - oauth
      summary: Get registered OAuth authenticators
      description: |
        Gets list of installed OAuth authenticators.
        
        Returns information about all configured OAuth providers (GitHub, GitLab, Bitbucket, etc.)
        that can be used for authentication.
        
        Each authenticator descriptor contains:
        - Provider name
        - Authentication endpoint URL
        - Links to authentication and token operations
      operationId: getRegisteredAuthenticators
      security:
        - BearerAuth: []
        - BasicAuth: []
      responses:
        '200':
          description: List of registered OAuth authenticators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OAuthAuthenticatorDescriptor'
              example:
                - name: github
                  endpointUrl: https://github.com/login/oauth/authorize
                  links:
                    - rel: authenticate
                      href: /oauth/authenticate?oauth_provider=github
                    - rel: token
                      href: /oauth/token?oauth_provider=github
                - name: gitlab
                  endpointUrl: https://gitlab.com/oauth/authorize
                  links:
                    - rel: authenticate
                      href: /oauth/authenticate?oauth_provider=gitlab
                    - rel: token
                      href: /oauth/token?oauth_provider=gitlab
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /oauth/token:
    get:
      tags:
        - oauth
      summary: Get OAuth token
      description: |
        Gets OAuth token for the authenticated user and specified provider.
        
        If no token exists, this endpoint will generate a new token (in demo mode)
        or trigger the OAuth flow (in production).
        
        The token can be used to access SCM provider APIs on behalf of the user.
      operationId: getOAuthToken
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - name: oauth_provider
          in: query
          description: OAuth provider name (e.g., github, gitlab, bitbucket)
          required: true
          schema:
            type: string
            enum: [github, gitlab, bitbucket, azure-devops]
          example: github
      responses:
        '200':
          description: OAuth token successfully retrieved or created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthToken'
              example:
                token: gho_1234567890abcdefghijklmnopqrstuvwxyz
                scope: repo user write:public_key
        '400':
          description: OAuth provider parameter is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Bad Request
                message: OAuth provider is required
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: OAuth provider not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Not Found
                message: OAuth provider not found
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    delete:
      tags:
        - oauth
      summary: Invalidate OAuth token
      description: |
        Invalidates (deletes) OAuth token for the authenticated user and specified provider.
        
        After invalidation, the user will need to re-authenticate with the OAuth provider
        to obtain a new token.
        
        This is useful for:
        - Revoking access to SCM providers
        - Security purposes (token rotation)
        - Testing OAuth flows
      operationId: invalidateOAuthToken
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - name: oauth_provider
          in: query
          description: OAuth provider name (e.g., github, gitlab, bitbucket)
          required: true
          schema:
            type: string
            enum: [github, gitlab, bitbucket, azure-devops]
          example: github
      responses:
        '204':
          description: OAuth token successfully invalidated (no content)
        '400':
          description: OAuth provider parameter is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Bad Request
                message: OAuth provider is required
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: OAuth token not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Not Found
                message: OAuth token not found
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    KubernetesNamespaceMeta:
      type: object
      required:
        - name
        - attributes
      properties:
        name:
          type: string
          description: |
            The name of the namespace.
            
            Value may be not a name of existing namespace, but predicted name 
            with placeholders inside, like `<workspaceid>`.
          example: che-johndoe
          pattern: '^[a-z0-9]([-a-z0-9]*[a-z0-9])?$'
          minLength: 1
          maxLength: 63
        attributes:
          type: object
          description: |
            Namespace attributes, which may contain additional info about the namespace.
            
            Common attributes:
            - `default`: "true" or "false" - shows if k8s namespace is configured as default
            - `phase`: "Active" or "Terminating" - current namespace status
          additionalProperties:
            type: string
          example:
            phase: Active
            default: 'true'
...
    OAuthToken:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: OAuth access token
          example: gho_1234567890abcdefghijklmnopqrstuvwxyz
        scope:
          type: string
          description: OAuth scope (space-separated list of permissions)
          example: repo user write:public_key
    OAuthAuthenticatorDescriptor:
      type: object
      required:
        - name
        - endpointUrl
      properties:
        name:
          type: string
          description: Name of the OAuth provider
          enum: [github, gitlab, bitbucket, azure-devops]
          example: github
        endpointUrl:
          type: string
          description: OAuth authorization endpoint URL
          example: https://github.com/login/oauth/authorize
        links:
          type: array
          description: HAL links for OAuth operations
          items:
            type: object
            properties:
              rel:
                type: string
                example: authenticate
              href:
                type: string
                example: /oauth/authenticate?oauth_provider=github
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type or category
          example: Unauthorized
        message:
          type: string
          description: Detailed error message
          example: Authorization header is required
        details:
          type: string
          description: Additional error details (only in development mode)
          example: Missing Authorization header in request
    ExtendedError:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            stack:
              type: string
              description: Stack trace (only in development mode)
              example: 'Error: Not able to find namespace\n    at NamespaceProvisioner.provision...'
  responses:
    Unauthorized:
      description: Unauthorized - Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missingAuth:
              value:
                error: Unauthorized
                message: Authorization header is required
            invalidAuth:
              value:
                error: Unauthorized
                message: Invalid authorization credentials
    InternalServerError:
      description: Internal server error occurred
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ExtendedError'
          examples:
            namespaceNotFound:
              value:
                error: Internal Server Error
                message: Not able to find namespace che-johndoe
            kubernetesError:
              value:
                error: Internal Server Error
                message: Failed to create namespace
                details: Kubernetes API error - Forbidden
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: userid:username
      description: |
        Bearer token authentication using the format `userid:username`.
        
        Example: `Authorization: Bearer user123:johndoe`
        
        In production, this should be replaced with proper JWT tokens.

    BasicAuth:
      type: http
      scheme: basic
      description: |
        Basic authentication using username and userid.
        
        - Username: user's username (e.g., `johndoe`)
        - Password: user's ID (e.g., `user123`)
        
        Example: `Authorization: Basic am9obmRvZTp1c2VyMTIz` (base64 of `johndoe:user123`)
security:
  - BearerAuth: []
  - BasicAuth: []